Cambios pequeños de codigo relacionados a Administrador

+++ src/main/java/pi/vetapp/repository/AdministradorRepository.java
@@ -5,7 +5,7 @@ import pi.vetapp.entity.Administrador;
 
 import java.util.Optional;
 
-public interface AdministradorRepository extends JpaRepository<Administrador, Integer> {
+public interface AdministradorRepository extends JpaRepository<Administrador, Long> {
   Optional<Administrador> findByUsuario(String usuario);
 
   Optional<Administrador> findByCorreo(String correo);
